+++ tests/components/advantage_air/test_climate.py
@@ -28,9 +28,22 @@ from homeassistant.helpers import entity_registry as er
 from . import add_mock_config, patch_get, patch_update
 
 
+@pytest.fixture
+def mock_get():
+    """Fixture to patch the Advantage Air async_get method."""
+    with patch_get() as mock_get:
+        yield mock_get
+
+
+@pytest.fixture
+def mock_update():
+    """Fixture to patch the Advantage Air async_get method."""
+    with patch_get() as mock_get:
+        yield mock_get
+
+
 async def test_climate_myzone_main(
-    hass: HomeAssistant,
-    entity_registry: er.EntityRegistry,
+    hass: HomeAssistant, entity_registry: er.EntityRegistry, mock_get
 ) -> None:
     """Test climate platform main entity."""
