+++ tests/components/advantage_air/test_init.py
@@ -1,10 +1,18 @@
 """Test the Advantage Air Initialization."""
 from advantage_air import ApiError
+import pytest
 
 from homeassistant.config_entries import ConfigEntryState
 from homeassistant.core import HomeAssistant
 
 from . import add_mock_config, patch_get
+
+
+@pytest.fixture
+def mock_get():
+    """Fixture to patch the Advantage Air async_get method."""
+    with patch_get() as mock_get:
+        yield mock_get
 
 
 async def test_async_setup_entry(hass: HomeAssistant, mock_get) -> None:
